Move map overlay components
