+++ src/containers/dashboard/map/Map.scss
@@ -16,13 +16,13 @@
 .DistrictMaps_Container {
   position: relative;
   flex: 3 0;
-  // z-index: 1;
+  z-index: 1;
 
   .react-transform-component,
   .react-transform-element {
     width: 100%;
     height: 100%;
-    // z-index: 1;
+    z-index: 1;
   }
   #flo-district-map {
     height: 100%;
@@ -35,5 +35,11 @@
   .distrcit:hover,
   .district:focus {
     fill: url(#diagonalHatchDark);
+  }
+
+  // TODO: WHY ISN'T THIS WORKING
+  .selected-district:hover,
+  .selected-district:focus {
+    stroke: 1px solid red;
   }
 }
